testing/web-platform/tests/atomics/window-and-any-worker/wake/wake-all-on-loc/wake-all-on-loc-popup.html

<!doctype html>
<meta charset=utf-8>
<title></title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<script>

// Copyright (C) 2017 Mozilla Corporation.  All rights reserved.
// This code is governed by the BSD license found in the LICENSE file.

/*---
description: >
  Test that Atomics.wake wakes all waiters on a location, but does not
  wake waiters on other locations.
---*/

window.addEventListener("load", function() {
  async_test(function(t) {
    let popup = window.open('../empty.html', 'empty');
    let test = this;
    popup.onmessage = function(e) {
      let rs = [];
      let workers = [];
      let nworkers = 4;

      // Create three workers and make them wait on zero.
      for (let i = 0; i < nworkers-1; i++) {
        let w = new Worker('../wait-on-zero.js');
        w.onmessage = function(e) {
          rs.push(e.data);
        }
        workers.push(w);
      }
      // Make fourth worker and make it wait on one.
      let w = new Worker('../wait-on-one.js');
      w.onmessage = function(e) {
        rs.push(e.data);
      }
      workers.push(w);
      // Send message to all workers.
      let ia = new Int32Array(new SharedArrayBuffer(2*Int32Array.BYTES_PER_ELEMENT));
      for (let i = 0; i < nworkers; i++) {
        workers[i].postMessage([ia.buffer, 2000]);
      }
      // Wake workers waiting on zero.
      setTimeout(() => assert_equals(3, Atomics.wake(ia, 0)), 500);
      // Check workers on zero woke but worker on one timed-out.
      setTimeout(test.step_func_done(function() {
        rs.sort();
        assert_equals(rs[0], "ok");
        assert_equals(rs[1], "ok");
        assert_equals(rs[2], "ok");
        assert_equals(rs[3], "timed-out");
      }), 2000);
    }
    popup.postMessage([], '*');
  }, "atomics-wake-all-on-loc-from-popup");
});
</script>
